# src/dovecord-react/src/services/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Dovecord API",
    "description": "Backend API for Dovecord",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:7045"
    }
  ],
  "paths": {
...
    "/api/v1/users/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "User_GetUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "User"
        ],
        "operationId": "User_DeleteUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "User"
        ],
        "operationId": "User_UpdateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "user",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserManipulationDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/users/me/channels": {
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "User_AddUserChannel",
        "requestBody": {
          "x-name": "recipientId",
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "format": "guid"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "UserDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "isOnline": {
            "type": "boolean",
            "nullable": true
          },
          "bot": {
            "type": "boolean"
          },
          "system": {
            "type": "boolean",
            "nullable": true
          },
          "accentColor": {
            "type": "boolean",
            "nullable": true
          }
        }
      },
      "ChannelDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "type": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "topic": {
            "type": "string",
            "nullable": true
          },
          "serverId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          }
        }
      },
...
      "UserManipulationDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "isOnline": {
            "type": "boolean",
            "nullable": true
          }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "oauth2",
        "description": "Type into the textbox: Bearer {your JWT token}.",
        "name": "Authorization",
        "scheme": "Bearer",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://danovas.b2clogin.com/danovas.onmicrosoft.com/B2C_1_signupsignin1/oauth2/v2.0/authorize",
            "tokenUrl": "https://danovas.b2clogin.com/danovas.onmicrosoft.com/B2C_1_signupsignin1/oauth2/v2.0/token",
            "scopes": {
              "https://danovas.onmicrosoft.com/89be5e10-1770-45d7-813a-d47242ae2163/API.Access": "Access the api as the signed-in user"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}